Add fizzBuzzWithoutDivMod.py file
This is unfortunately not original work. My initial assumption was run an inner loop to check the number if it fits the fizz, buzz or fizzbuzz criteria. Though I instantly know that it is not a good implementation as it would have a quadratic big O notation (O)n^2 due to the two nested loops. I tried to apply the saying "if in doubt, throw a hashmap (dictionary) at the problem." which I think is a bit overkill. So in the end I looked up online and their implementation involves setting up two variables that simple counts to the needed criteria. It does make sense, since counting would inevitably traverse numbers that are multiples of 3 or 5. 

If no criteria is met, a flag variable is checked used to print the number. It is understood that division and modulus division are computationally expensive operations, especially as the numbers grow bigger and bigger. In computer science it is expected that programs may accept numbers or arrays of infinite sizes. Therefore ensuring algorithm efficiency is crucial. I definitely learned something interesting from this, I hope that I can do more with programming in the future.